+++ cmd/hlt/main.go
@@ -35,8 +35,7 @@ func main() {
 	args := os.Args[1:]
 
 	if ok := gFlags.Validate(args); !ok {
-		log.Fatal("args invalid")
-		return
+		panic("args invalid")
 	}
 
 	if gFlags.Get("version").GetBoolValue(args) {